Fix generator invalidation with decorator
My wrappers over MorphoDiTa tokenizers and taggers create generators which rely
on shared state -- specifically, all of these generators have the same tokenizer
and setting a new text for the tokenizer in a new generator affects all
previously created generators. That's bad.

This commit solves the issue by introducing a decorator for the instance methods
creating those generators which truncates any previously created generators when
a new one is created, and printing a warning.

However, this might not be the best way to do it. After thinking about it, the
only problematic part is the sharing of the underlying tokenizer, and tokenizers
might be cheap enough to create and keep around (unlike underlying taggers) that
it might make sense to just create a separate tokenizer for each generator and
be done with it. The behavior will be fully intuitive, and without the need for
truncating, decorators and printing any warnings.