Fix generator invalidation by only sharing taggers
This is a much better approach -- MorphoDiTa taggers are shared, because they're
expensive to load (and possibly large?), but all other objects are created for
each generator separately, which means that if multiple generators exist for a
single tagger (or tokenizer), they don't stomp on each other's state.

Sharing the underlying tagger is not a problem because it doesn't keep around
any state which should persist between calls to the `.tag()` method, and the
`.tag()` method itself can only be called by one of the generators which share
the tagger at a time.