Make API docs build on ReadTheDocs
The API docs are built with autodoc, which means corpy needs to be installed on
ReadTheDocs. However, the resource limits on ReadTheDocs make it impossible to
do a full install of corpy with all dependencies, so I had to work around this
by introducing a dummy setup.py which is only used to install corpy on
ReadTheDocs and omits most of the dependencies, leaving it up to Sphinx to mock
them.