Fix documentation builds
Also, make the configuration simpler. The builds used to time out because
installation of dependencies took too long, so there was a dummy setup.py and
MANIFEST.in with "light" deps for the sole benefit of ReadTheDocs. This setup
broke down for some reason (well it was very fragile and complicated in the
first place...), but just pip-installing corpy seems not to be timing out
anymore, and seems to be the recommended way to do things. So let's go with
that and see how it works.

Note that there are still some dev dependencies that need to be installed
separately, but these are now kept in docs/requirements.txt, which is generated
by check.sh and only used by ReadTheDocs.