clean_env: only affect function calls if requested
After considering various sorts of source code manipulation (via AST), which
would probably restrict this functionality to the IPython magic only (which has
easy access to the source code), it turns out clever use of tracing is probably
the best and easiest (though still tricky) solution. Plus it makes the
functionality available across the board, not only in IPython.