IPython magic for clean_env, attempt #1
This is as far as I got trying to have the `with clean_env(...)` context manager
in the magic function, instead of injecting it into the cell. Strict mode works
fine, you just give it the env to clean up, but non-strict mode is tricky. I'd
have to figure out when exactly to start tracing and when to stop, so that I
don't accidentally break anything but the code in the user's cell (which I am
supposed to break).

I went so far as to enable/disable tracing in pre/post_execute callbacks (cf.
https://ipython.readthedocs.io/en/stable/config/callbacks.html), but that still
breaks some layers between these sites and actually running the user code (as
can be seen by printing the stack frames in the global_trace callback). Figuring
out the precise places where I need to enable/disable tracing, and how to hook
into those, is probably way too much work and way too hacky to get working; at
this point, inserting the with-statement into the cell is probably less finicky.