IPython magic for clean_env, attempt #2
It's probably best to make the magic command work basically like a macro, just
syntactic sugar for writing out the with block by hand. It guarantees you get
the same behavior without having to jump through a lot of hoops specific to the
situation when you run it as a magic, which can be brittle. Plus having the full
clean_env invocation spelled out in the tracebacks if anything goes wrong can
actually be helpful.