Finalize transition to src/ subdir
Poetry supports the src/ layout without the need for any configuration. To the
contrary: trying to explicitly add the package from the src/ subdir breaks
everything (Poetry then fails to install the package into the virtualenv, which
breaks tests etc.).

Also, I'd failed to update the paths the linters are pointed at.