Remove util.{head,cmp}
head was motivated by NLTK suddenly spewing long lists of tokens instead of
previews, as it'd been doing previously, but I never remember to use it anyway,
and it's better to use the occasion to remind students about slicing. Also, it
feels like NLTK has fixed this in some places, reverting to the old behavior
(just a hunch based, I feel like I've been less frustrated with this issue over
recent iterations of my course).

As for cmp -- again, I never remember to use it, it's somewhat brittle (because
it relies on head, among other things), and it's probably better to write
asserts so that 1) students can learn about them, and 2) you can have custom
behavior and messages more appropriate in the given context.

Closes #7.