Smarter treatment of vowel sequences
Mostly by using the lemmatization and morphological derivation provided
by a MorphoDiTa tagger, with some exceptions (as in, from
exceptions.tsv) sprinkled on top.

At this point, it's mostly about what (not to) turn into a /*_u/
diphthong and where (not to) insert a /j/ on hiatus, but in theory, it
could be extended e.g. to insert plausible glottal stops.

This commit also updates many regression tests which were wrong in this
respect (they contained a diphthong when they shouldn't have). I seem to
remember this was intentional at the time (a known bug/compromise), but
still, this update is a net improvement.

Fixes #10.